Input Handler Component, Sprite Renderer Component, Minor Updates
Input Handler:
Uses a map to store a pointer to a function using an keyboard key as the
key.  Iterattes thru them to find if a key was pressed and if so run the
corresponding function.  (needs some work as of right now can only take
in functions that arent tied to a class)
Sprite Renderer:
Can load sprites from files
Renders sprites on update
(Add error if sprite is empty)
Minor Updates:
Components now have pointer pointing to the attached game object.  This
pointer is set up on game object adding said component.