Gameobject get component, Box Collider, Minor changes
Gameobject Get Component:
Now works, returns a pointer to the component given in the function
template, returns null pointer if the component cannot be found (only
works on the first instance of the component as intended)
Box Collider:
Contains half width and height, has bottom left and top right variables.
These are used to calculate collisions. NOTE: object must have rigid
body attached to work properly. Also this is not yet connected to the
rigid body component so effectivly does nothing as of yet.
Minor changes:
Made delta_time float in main loop that is set to the
restart.asmiliseconds() of the clock variable to save cpu time.