Sound and music Compoents, Debug Class
Sound and Music Components:
Added sound for sfx and music for music.
Can start/stop/pause components.
Debug Class:
Allows you to print/clear/log error to console with option to halt game
engine. (needs work because it halts the thread and the engine only runs
on one thread so windows will think that the window is not responding)