Scene, Gameobject and Component (Final)
Scene:
Added scene objects via XML
Added lander locations so the player can win
Added a menu screen that lets the player know if they won or died.
Gameobject and Component:
Added a property called enabled for both these which does a check if is
true and if it isnt then it doesnt proceed with start / update.  Can be
toggled to turn on / off gameobjects objects.